requirements.txt
# Common
joblib==0.11
PyYAML==3.12
termcolor==1.1.0

# Twine
twine==1.13.0
# - (deps)
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
docutils==0.15.2
idna==2.8
pkginfo==1.5.0.1
Pygments==2.4.2
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
tqdm==4.32.2
urllib3==1.25.3
webencodings==0.5.1

# PyTest
pytest==5.0.1
# - (deps)
atomicwrites==1.3.0
attrs==19.1.0
importlib-metadata==0.19
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.2
PyYAML==3.12
wcwidth==0.1.7
zipp==0.5.2
